+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 7825f61 - 2018-05-05: Pre-Release v9.3.20.dev3 
+- * 75b93e9 - 2018-05-06: Fixed UnicodeEncode errors when saving subtitles 
+- * f69a02b - 2018-05-05: Pre-Release v9.3.20.dev3 
 - * b7cc531 - 2018-05-05: Refactoring search client code 
 - * e0d01bd - 2018-05-05: Fixed redirect on login Changed secure cookie variable name 
 - * b5b3a06 - 2018-05-05: Pre-Release v9.3.20.dev2